R-Link-Server/requirements.txt
# R-Link-Server 依赖

# Web 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 进程管理
psutil>=5.9.0

# 配置文件解析
pyyaml>=6.0

# 数据验证
pydantic>=2.0.0

# CORS
aiohttp>=3.9.0
